Split UI build, optimize with no WASM
