+++ package.json
@@ -19,8 +19,8 @@
     "typescript": "^4.1.3"
   },
   "scripts": {
-    "build": "yarn build:i18n && yarn build:code && yarn build:zip",
-    "build:code": "NODE_ENV=production polkadot-dev-build-ts",
+    "build": "polkadot-dev-build-ts && yarn build:zip",
+    "build:extra": "yarn build:i18n && (cd packages/extension && yarn webpack --config webpack.extension.js)",
     "build:i18n": "i18next-scanner --config i18next-scanner.config.js",
     "build:release": "polkadot-ci-ghact-build",
     "build:zip": "yarn build:zip:dst && yarn build:zip:src",
